refactor components for one united design
