+++ SportClubsFrontend/src/App.tsx
@@ -3,6 +3,7 @@ import { Routes, Route } from 'react-router-dom';
 import SignIn from './components/SignIn';
 import ResetForm from './components/ResetForm';
 import ClubCreationForm from './components/ClubCreationForm';
+import AthleteForm from './components/AthleteForm';
