+++ SportClubsFrontend/src/styles/modalStyles.tsx
@@ -1,0 +1,12 @@
+export const sxCard = {
+  boxShadow: '0 4px 16px rgba(0, 0, 0, 0.1)',
+  padding: '15px',
+  marginTop: 8,
+  borderRadius: 5,
+};
+
+export const sxDialog = {
+  style: {
+    borderRadius: 20,
+  },
+};
