+++ SportClubsFrontend/src/styles/styles.tsx
@@ -1,17 +1,14 @@
 import { createTheme } from '@mui/material/styles';
 
 const Colors = {
-    primary: '#4741DE',
-       
-}
-
-
+  primary: '#4741DE',
+};
 const theme = createTheme({
-    palette:{
-        primary: {
-            main: Colors.primary
-        }
-    }
-})
+  palette: {
+    primary: {
+      main: Colors.primary,
+    },
+  },
+});
 
-export default theme
+export default theme;
